+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - Add `funding_source_description` to PayPalAccount
 - Add missing `GooglePayCard` error code
 - Add `REFUND_FAILED` to Webhook Notification
+- Add `final_capture` to Transaction `_submit_for_partial_settlement_signature`
 
 ## 4.20.0